feat: export default helia config
To make it easier to customise the default Helia settings, export
a `heliaDefaults` function similar to the existing
`libp2pDefaults` function.
